+++ opentelemetry-otlp/src/exporter/http/metrics.rs
@@ -10,14 +10,12 @@ use super::OtlpHttpClient;
 
 impl MetricsClient for OtlpHttpClient {
     async fn export(&self, metrics: &ResourceMetrics) -> OTelSdkResult {
-        let build_body_wrapper = |client: &OtlpHttpClient, metrics: &ResourceMetrics| {
-            client
-                .build_metrics_export_body(metrics)
-                .ok_or_else(|| "Failed to serialize metrics".to_string())
-        };
-
         let response_body = self
-            .export_http_with_retry(metrics, build_body_wrapper, "HttpMetricsClient.Export")
+            .export_http_with_retry(
+                metrics,
+                OtlpHttpClient::build_metrics_export_body,
+                "HttpMetricsClient.Export",
+            )
             .await?;
 
         handle_partial_success(&response_body, self.protocol);