+++ opentelemetry-stdout/src/metrics/exporter.rs
@@ -1,5 +1,5 @@
 use chrono::{DateTime, Utc};
-use core::{f64, fmt};
+use core::fmt;
 use opentelemetry_sdk::metrics::data::{AggregatedMetrics, MetricData};
 use opentelemetry_sdk::metrics::Temporality;
 use opentelemetry_sdk::{